Move the location to another contact link
This has the added benefit of making the flexbox look cleaner, although
maybe it should be a grid instead... Gonna think on it.

TODO: There's no way with OpenStreetMap to link directly to a city with only its
name, so the link only goes to a search page. I should query the OSM API and
get the correct link for the city.